+++ projects/story-decorator/decorator.ts
@@ -10,109 +10,98 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { html, render, TemplateResult } from '@spectrum-web-components/base';
+import { html } from '@spectrum-web-components/base';
+import { useEffect } from '@storybook/preview-api';
 import './sp-story-decorator.js';
-import { Parameters, Renderer, StoryContext } from '@storybook/csf';
+import type { StoryContext, StoryFn } from '@storybook/web-components';
 
 export const themeStyles = html`
     <style>
         #root {
             padding: 0;
         }
-        .docs-story sp-story-decorator::part(container) {
+        sp-story-decorator::part(container) {
             min-height: auto;
             position: relative;
         }
-        .docs-story sp-story-decorator::part(controls) {
+        sp-story-decorator::part(controls) {
             position: absolute;
         }
     </style>
 `;
 
-export const swcThemeDecoratorWithConfig =
-    ({ bundled } = { bundled: true }) =>
-    (
-        story: () => TemplateResult,
-        context: StoryContext<Renderer, Parameters>
-    ) => {
-        // Needed so that the `sp-language-context` event can be caught by decorator's `sp-theme`
-        requestAnimationFrame(() => {
-            document.documentElement.setAttribute('lang', 'en-US');
-            const decorator = document.querySelector(
-                'sp-story-decorator'
-            ) as HTMLElement;
-            render(bundled ? story() : html``, decorator);
-        });
+/**
+ * Global properties added to each component; determines what stylesheets are loaded
+ **/
+export const swcThemeDecorator = (story: StoryFn, context: StoryContext) => {
+    const {
+        globals: {
+            system,
+            color,
+            scale,
+            textDirection,
+            reduceMotion,
+            lang,
+        } = {},
+    } = context;
 
-        const isGlobalSettingDefined = (setting: string): boolean => {
-            return context?.globals?.[setting] !== undefined;
-        };
-
-        const globalSettings = [
-            'system',
-            'color',
-            'scale',
-            'textDirection',
-            'reducedMotion',
-            'locale',
-        ];
-        let hideNavStyles;
-
-        if (globalSettings.every(isGlobalSettingDefined)) {
-            hideNavStyles = html`
-                <style>
-                    sp-story-decorator::part(controls) {
-                        display: none;
-                    }
-                </style>
-            `;
-        }
-
+    useEffect(() => {
         // Update window.__swc_hack_knobs__ values with current context globals
-        if (context?.globals) {
-            if (context.globals.system) {
-                window.__swc_hack_knobs__.defaultSystemVariant =
-                    context.globals.system;
-            }
-            if (context.globals.color) {
-                window.__swc_hack_knobs__.defaultColor = context.globals.color;
-            }
-            if (context.globals.scale) {
-                window.__swc_hack_knobs__.defaultScale = context.globals.scale;
-            }
-            if (context.globals.textDirection) {
-                window.__swc_hack_knobs__.defaultDirection =
-                    context.globals.textDirection;
-                if (
-                    document.documentElement.dir !==
-                    context.globals.textDirection
-                ) {
-                    document.documentElement.dir =
-                        context.globals.textDirection;
-                }
-            }
-            if (context.globals.reduceMotion !== undefined) {
-                window.__swc_hack_knobs__.defaultReduceMotion =
-                    context.globals.reduceMotion;
-            }
-            if (context.globals.lang) {
-                window.__swc_hack_knobs__.defaultLocale = context.globals.lang;
+        if (system) {
+            window.__swc_hack_knobs__.defaultSystemVariant = system;
+        }
+        if (color) {
+            window.__swc_hack_knobs__.defaultColor = color;
+        }
+        if (scale) {
+            window.__swc_hack_knobs__.defaultScale = scale;
+        }
+        if (textDirection) {
+            window.__swc_hack_knobs__.defaultDirection = textDirection;
+            if (document.documentElement.dir !== textDirection) {
+                document.documentElement.dir = textDirection;
             }
         }
+        if (reduceMotion !== undefined) {
+            window.__swc_hack_knobs__.defaultReduceMotion = reduceMotion;
+        }
+        if (lang) {
+            window.__swc_hack_knobs__.defaultLocale = lang;
+        }
+    }, [system, color, scale, textDirection, reduceMotion, lang]);
 
-        return html`
-            ${themeStyles} ${hideNavStyles}
-            <sp-story-decorator
-                role="main"
-                system=${context?.globals?.system}
-                color=${context?.globals?.color}
-                scale=${context?.globals?.scale}
-                locale=${context?.globals?.locale}
-                lang=${context?.globals?.lang}
-                .direction=${context?.globals?.textDirection}
-                ?reduce-motion=${context?.globals?.reduceMotion}
-            ></sp-story-decorator>
-        `;
-    };
+    const hasAnySetting =
+        system || color || scale || textDirection || reduceMotion;
 
-export const swcThemeDecorator = swcThemeDecoratorWithConfig();
+    return html`
+        <style>
+            #root {
+                padding: 0;
+            }
+            sp-story-decorator::part(container) {
+                min-height: auto;
+                position: relative;
+            }
+            sp-story-decorator::part(controls) {
+                position: absolute;
+            }
+            ${hasAnySetting
+                ? `sp-story-decorator::part(controls) {
+                display: none;
+            }
+        `
+                : ''}
+        </style>
+        <sp-story-decorator
+            role="main"
+            system=${system}
+            color=${color}
+            scale=${scale}
+            lang=${lang}
+            .direction=${textDirection}
+            ?reduce-motion=${reduceMotion}
+        >
+            ${story({}, context)}
+        </sp-story-decorator>
+    `;
+};